pyproject.toml
[tool.poetry]
name = "compliant-cdk-lib"
version = "0.1.3"
description = ""
authors = ["Keith Lagrange <punolagrange@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "2.153"
pytest = "^8.3.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
